Moves extension loader into separate file
