+++ c/meterpreter/workspace/metsrv/metsrv.vcxproj.filters
@@ -30,6 +30,7 @@
     <ClInclude Include="..\..\source\metsrv\server_http_utils.h" />
     <ClInclude Include="..\..\source\metsrv\winapi.h" />
     <ClInclude Include="..\..\source\metsrv\rc4.h" />
+    <ClInclude Include="..\..\source\metsrv\extension_loader.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\source\metsrv\pivot_tree.c" />
@@ -63,6 +64,7 @@
     <ClCompile Include="..\..\source\metsrv\server_http_utils.c" />
     <ClCompile Include="..\..\source\metsrv\winapi.c" />
     <ClCompile Include="..\..\source\metsrv\rc4.c" />
+    <ClCompile Include="..\..\source\metsrv\extension_loader.c" />
   </ItemGroup>
   <ItemGroup>
     <MASM Include="..\..\source\ReflectiveDLLInjection\dll\src\GateTrampoline64.asm" />
